Add support for DBLP.
This code uses DBLP's XML-based query interface, rather than HTML
screen-scraping.  The advantage of using XML is that we can parse results
more reliably, without breaking if DBLP decides to change the look of their
web pages.  The disadvantage of using XML is that we require Emacs to have
been compiled with libxml2 support so that we have access to a real XML
parser.