Fabricate a cookie to inject for url-retrieve.
Unfortunately, adding a cookie as an extra header ends up with multiple
cookie headers, which isn't really allowed in current HTTP
specifications, and not accepted by Google in any case. (There is a
comment in url-cookie.el complaining about servers that don't accept
multiple Cookie: headers, but the current specs are clear on this now.)

So we inject a new cookie into the Emacs cookie storage, set to expire
an hour from now.