Clean some punctuation from queries.
Some punctuation in a query can confuse the lookups on Google Scholar.
Such punctuation is unusual (but possible) in queries typed by a user,
but it is more likely in machine-generated queries.

Also, to format the query string, query terms need to be separated by
+ symbols. However, url-hexify-string hex-encodes the + symbols, so
they are not interpreted as separators by the server. So we hexify
first, then replace hex-encoded space characters by + in the query.