Use window quit-restore for gscholar and entry window
quit-restore-window handles killing or keeping the window used by gscholar-buffer and entry-buffer based on the history of buffers previously displayed in the window. Ie. the window is killed if it was opened just for gscholar, but preserved if it existed already.
As split-window-below used for opening the entry-window messes with the buffer history used by quit-restore-window, it is manually adapted as discussed in https://lists.gnu.org/archive/html/help-gnu-emacs/2017-03/msg00006.html